+++ auth/browser.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/rusq/slackdump/v3/auth/auth_ui"
 	"github.com/rusq/slackdump/v3/auth/browser"
+	"github.com/rusq/slackdump/v3/internal/structures"
 )
 
 var _ Provider = BrowserAuth{}
@@ -58,7 +59,7 @@ func NewBrowserAuth(ctx context.Context, opts ...Option) (BrowserAuth, error) {
 		}
 		defer br.opts.flow.Stop()
 	}
-	if wsp, err := auth_ui.Sanitize(br.opts.workspace); err != nil {
+	if wsp, err := structures.ExtractWorkspace(br.opts.workspace); err != nil {
 		return br, err
 	} else {
 		br.opts.workspace = wsp
