+++ internal/structures/structures.go
@@ -3,7 +3,10 @@ package structures
 
 import (
 	"errors"
+	"fmt"
+	"net/url"
 	"regexp"
+	"strings"
 )
 
 const (
@@ -25,4 +28,29 @@ func ValidateToken(token string) error {
 		return errInvalidToken
 	}
 	return nil
+}
+
+var ErrInvalidDomain = errors.New("invalid domain")
+
+// ExtractWorkspace takes a workspace name or URL and returns the workspace name.
+func ExtractWorkspace(workspace string) (string, error) {
+	if !strings.Contains(workspace, ".slack.com") && !strings.Contains(workspace, ".") {
+		return workspace, nil
+	}
+	if strings.HasPrefix(workspace, "https://") {
+		uri, err := url.Parse(workspace)
+		if err != nil {
+			return "", err
+		}
+		workspace = uri.Host
+	}
+	// parse
+	name, domain, found := strings.Cut(workspace, ".")
+	if !found {
+		return "", errors.New("workspace name is empty")
+	}
+	if strings.TrimRight(domain, "/") != "slack.com" {
+		return "", fmt.Errorf("%s: %w", domain, ErrInvalidDomain)
+	}
+	return name, nil
 }
